🧹 update equinix provider to support both org and project
