+++ providers/equinix/resources/equinix.lr
@@ -5,21 +5,19 @@ option provider = "go.mondoo.com/cnquery/providers/equinix"
 option go_package = "go.mondoo.com/cnquery/providers/equinix/resources"
 
 // Equinix Metal project
-equinix.metal.project {
+equinix.metal.project @defaults("name") {
   id string
   name string
   organization() equinix.metal.organization
   createdAt time
   updatedAt time
   url string
-  paymentMethod dict
-  users() []equinix.metal.user
   sshKeys() []equinix.metal.sshkey
   devices() []equinix.metal.device
 }
 
 // Equinix Metal organization
-equinix.metal.organization {
+equinix.metal.organization @defaults("name") {
   id string
   name string
   description string
@@ -33,10 +31,11 @@ equinix.metal.organization {
   billingPhone string
   creditAmount float
   url string
+ 	users() []equinix.metal.user
 }
 
 // Equinix Metal user
-equinix.metal.user {
+private equinix.metal.user @defaults("email") {
   id string
   firstName string
   lastName string
@@ -52,11 +51,10 @@ equinix.metal.user {
   timezone string
   phoneNumber string
   url string
-  vpn bool
 }
 
 // Equinix Metal SSH key
-equinix.metal.sshkey {
+equinix.metal.sshkey @defaults("label") {
   id string
   label string
   key string
